Client Player movement test
Updated the Client player and Player to now work properly, allowing a demonstration of the movement planned for the game. Several things have been move to the client for testing purposes, with some hard coding. These will be removed soon, at the same time as moving things back to player. Changes made are for the demonstration only, and is mainly a result of problems with the base.Update() function